Move collection view protocols to another file
