+++ PhotoApp/PhotoApp/CollectionViewDataSource.swift
@@ -1,0 +1,26 @@
+//
+//  CollectionViewDataSource.swift
+//  PhotoApp
+//
+//  Created by 조중윤 on 2021/03/23.
+//
+
+import UIKit
+
+class CollectionViewDataSource: NSObject, UICollectionViewDataSource {
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return 40
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: "PhotoCell", for: indexPath) as! PhotoCell
+        return cell
+    }
+}
+
+extension CollectionViewDataSource: UICollectionViewDelegateFlowLayout {
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        return CGSize(width: 100.0, height: 100.0)
+    }
+}
+
